Extract RequestTemplateFactoryResolver
